Restructure as a Swift Package
I realized by the end of chapter 2 that the structure of the programs for this book will not be full macOS apps, but simple executables. Therefore this project does not need the macOS app target, it needs multiple executables. I tried redoing this as an Xcode project with multiple command line tools, only to find out that Xcode's XCTest doesn't work well with command line tools. I could only get XCTest to work with an app target in the Xcode project.

So I tried a Swift Package. With this structure I'm able to have multiple executable targets, and use XCTest.